Reorganize directory and namespace
